requirements.txt
streamlit
requests
pillow
python-dotenv

# Development tools (optional)
# black>=23.0.0
# isort>=5.12.0
# pylint>=2.17.0
# mypy>=1.0.0
